Refac. moved tests to relevant projects.
