+++ source/dotnet/Services/ProcessManager.IntegrationTests/TestHelpers/ServiceProviderHelpers.cs
@@ -14,7 +14,7 @@
 
 using Energinet.DataHub.Wholesale.IntegrationTests.TestHelpers;
 
-namespace Energinet.DataHub.Wholesale.IntegrationTests.Fixtures.TestHelpers;
+namespace ProcessManager.IntegrationTests.TestHelpers;
 
 internal static class ServiceProviderHelpers
 {
